Sophon

Documentation for Sophon.

Sophon.PINNAttentionNet

Sophon.TriplewiseFusion

Sophon.PINNAttentionNet — Type

PINNAttentionNet(H_net, U_net, V_net, fusion::TriplewiseFusion)
PINNAttentionNet(H_net, U_net, V_net; fusion_layers)

The output dimesion of H_net and the input dimension of layers must be the same .

           x → U_net → u                     u
                         ↘                     ↘
x → H_net →  h1 → layer1 → connection → layer2 → connection
                         ↗                     ↗
           x → V_net → v                     v

References

[1] Wang, Sifan, Yujun Teng, and Paris Perdikaris. "Understanding and mitigating gradient flow pathologies in physics-informed neural networks." SIAM Journal on Scientific Computing 43.5 (2021): A3055-A3081

Sophon.TriplewiseFusion — Type

TriplewiseFusion(connection, layers...)

         u1                    u2
            ↘                     ↘
h1 → layer1 → connection → layer2 → connection
            ↗                     ↗
         v1                    v2

Arguments

connection: Takes 3 inputs and combines them

layers: AbstractExplicitLayers

Inputs

Layer behaves differently based on input type:

A tripe of (h, u, v), where u and v itself are tuples of length N, the layers is also a tuple of length N. The computation is as follows

for i in 1:N
    h = connection(layers[i](h), u[i], v[i])
end

A triple of (h, u, v), where u and v are AbstractArrays.

for i in 1:N
    h = connection(layers[i](h), u, v)
end
h = layers[end](h)

Returns

See Inputs section for how the return value is computed

Updated model state for all the contained layers

Parameters

Parameters of each layer wrapped in a NamedTuple with fields = layer_1, layer_2, ..., layer_N

States

States of each layer wrapped in a NamedTuple with fields = layer_1, layer_2, ..., layer_N