Sophon

Documentation for Sophon.

Sophon.FourierFeature

Sophon.MultiscaleFourier

Sophon.PINNAttention

Sophon.TriplewiseFusion

Sophon.FourierFeature — Type

FourierFeature(in_dims::Int; num_modes::Int, std::Number=10.0f0)
FourierFeature(int_dims::Int, out_dims::Int; std::Number=10.0f0)

Fourier Feature Network.

Keyword Arguments

num_modes: Number of modes.

std: Standard deviation of the Gaussian distribution from which the frequencies are sampled.

States

modes: Random Fourier mappings.

Sophon.MultiscaleFourier — Type

MultiscaleFourier(int_dims::Int, out_dims::Int, hidden_dims::Int, chain; std)

Multi-scale Fourier Net.

    FourierFeature → chain_1
  ↗                          ↘
x → FourierFeature → chain_2 → Dense → y
  ↘                          ↗
    FourierFeature → chain_3

Here chain_1, chain_2, and chain_3 are all identical copies of chain, but with different parameters.

Keyword Arguments

std: A container of standard deviations of the Gaussian distribution.

References

[1] Wang, Sifan, Hanwen Wang, and Paris Perdikaris. “On the eigenvector bias of fourier feature networks: From regression to solving multi-scale pdes with physics-informed neural networks.” Computer Methods in Applied Mechanics and Engineering 384 (2021): 113938.

Sophon.PINNAttention — Type

PINNAttention(H_net, U_net, V_net, fusion::TriplewiseFusion)
PINNAttention(H_net, U_net, V_net; fusion_layers)

The output dimesion of H_net and the input dimension of layers must be the same .

           x → U_net → u                     u
                         ↘                     ↘
x → H_net →  h1 → layer1 → connection → layer2 → connection
                         ↗                     ↗
           x → V_net → v                     v

References

[1] Wang, Sifan, Yujun Teng, and Paris Perdikaris. "Understanding and mitigating gradient flow pathologies in physics-informed neural networks." SIAM Journal on Scientific Computing 43.5 (2021): A3055-A3081

Sophon.TriplewiseFusion — Type

TriplewiseFusion(connection, layers...)

         u1                    u2
            ↘                     ↘
h1 → layer1 → connection → layer2 → connection
            ↗                     ↗
         v1                    v2

Arguments

connection: Takes 3 inputs and combines them

layers: AbstractExplicitLayers

Inputs

Layer behaves differently based on input type:

A tripe of (h, u, v), where u and v itself are tuples of length N, the layers is also a tuple of length N. The computation is as follows

for i in 1:N
    h = connection(layers[i](h), u[i], v[i])
end

A triple of (h, u, v), where u and v are AbstractArrays.

for i in 1:N
    h = connection(layers[i](h), u, v)
end

Returns

See Inputs section for how the return value is computed

Updated model state for all the contained layers

Parameters

Parameters of each layer wrapped in a NamedTuple with fields = layer_1, layer_2, ..., layer_N

States

States of each layer wrapped in a NamedTuple with fields = layer_1, layer_2, ..., layer_N