Sophon

Documentation for Sophon.

Sophon.FourierFeature

Sophon.PINNAttention

Sophon.RBF

Sophon.Sine

Sophon.TriplewiseFusion

Sophon.FourierAttention

Sophon.FullyConnected

Sophon.MultiscaleFourier

Sophon.Siren

Sophon.SirenAttention

Sophon.gaussian

Sophon.kaiming_normal

Sophon.kaiming_uniform

Sophon.wu

Sophon.FourierFeature — Type

FourierFeature(in_dims::Int, modes::NTuple{N,Pair{S,T}}) where {N,S,T<:Int}

Fourier Feature Network.

\[\phi^{(i)}(x)=\left[\sin \left(2 \pi W^{(i)} x\right) ; \cos 2 \pi W^{(i)} x\right], \quad W^{(i)} \sim \mathcal{N}\left(0, \sigma^{(i)}\right)\]

Arguments

in_dims: Number of the input dimensions.

modes: A tuple of pairs of std => out_dims, where std is the standard deviation of the Gaussian distribution, and out_dims is the corresponding number of output dimensions.

Inputs

x: An an AbstractArray with size(x, 1) == in_dims.

Returns

An AbstractArray with size(y, 1) == sum(last(modes) * 2).

The states of the modes.

States

States of each mode wrapped in a NamedTuple with fields = mode_1, mode_2, ..., mode_N.

Examples

julia> f = FourierFeature(2, (1 => 3, 50 => 4))
FourierFeature(2 => 14)

julia> ps, st = Lux.setup(rng, f)
(NamedTuple(), (mode_1 = Float32[0.7510394 0.0678698; -1.6466209 -0.08511321; -0.4704813 2.0663197], mode_2 = Float32[-98.90031 -42.593884; 110.35572 15.565719; 81.60114 51.257904; -0.53021294 15.216658]))

julia> st
(mode_1 = Float32[0.7510394 0.0678698; -1.6466209 -0.08511321; -0.4704813 2.0663197], mode_2 = Float32[-98.90031 -42.593884; 110.35572 15.565719; 81.60114 51.257904; -0.53021294 15.216658])

References

[1] Tancik, Matthew, et al. “Fourier features let networks learn high frequency functions in low dimensional domains.” Advances in Neural Information Processing Systems 33 (2020): 7537-7547.

Sophon.PINNAttention — Type

PINNAttention(H_net, U_net, V_net, fusion_layers)
PINNAttention(in_dims::Int, out_dims::Int, activation::Function=sin;
              hidden_dims::Int, num_layers::Int)

The output dimesion of H_net and the input dimension of fusion_layers must be the same. For the second and the third constructor, Dense layers is used for H_net, U_net, and V_net. Note that the first constructer does not contain the output layer.

                 x → U_net → u                           u
                               ↘                           ↘
x → H_net →  h1 → fusionlayer1 → connection → fusionlayer2 → connection
                               ↗                           ↗
                 x → V_net → v                           v

Arguments

- `H_net`: `AbstractExplicitLayer`.
- `U_net`: `AbstractExplicitLayer`.
- `V_net`: `AbstractExplicitLayer`.
- `fusion_layers`: `Chain`.

Keyword Arguments

num_layers: The number of hidden layers.

hidden_dims: The number of hidden dimensions of each hidden layer.

References

[1] Wang, Sifan, Yujun Teng, and Paris Perdikaris. "Understanding and mitigating gradient flow pathologies in physics-informed neural networks." SIAM Journal on Scientific Computing 43.5 (2021): A3055-A3081

Sophon.RBF — Type

RBF(in_dims::Int, out_dims::Int, num_centers::Int=out_dims; sigma::AbstractFloat=0.2f0)

Radial Basis Fuction Network.

Sophon.Sine — Type

Sine(in_dims::Int, out_dims::Int, activation=sin; is_first::Bool = false, omega::AbstractFloat = 30f0)

Sinusoidal layer.

Example

s = Sine(2, 2; is_first=true) # first layer
s = Sine(2, 2) # hidden layer
s = Sine(2, 2, identity) # last layer

Sophon.TriplewiseFusion — Type

TriplewiseFusion(connection, layers...)

         u1                    u2
            ↘                     ↘
h1 → layer1 → connection → layer2 → connection
            ↗                     ↗
         v1                    v2

Arguments

connection: A functio takes 3 inputs and combines them.

layers: AbstractExplicitLayers or a Chain.

Inputs

Layer behaves differently based on input type:

A tripe of (h, u, v), where u and v itself are tuples of length N, the layers is also a tuple of length N. The computation is as follows

for i in 1:N
    h = connection(layers[i](h), u[i], v[i])
end

A triple of (h, u, v), where u and v are AbstractArrays.

for i in 1:N
    h = connection(layers[i](h), u, v)
end

Parameters

Parameters of each layer wrapped in a NamedTuple with fields = layer_1, layer_2, ..., layer_N

States

States of each layer wrapped in a NamedTuple with fields = layer_1, layer_2, ..., layer_N

Sophon.FourierAttention — Function

FourierAttention(in_dims::Int, out_dims::Int, activation::Function=sin;
                 hidden_dims::Int=512, num_layers::Int=6, modes::NTuple)

x → [FourierFeature(x); x] → PINNAttention

Arguments

in_dims: The input dimension.

Keyword Arguments

modes: A tuple of pairs of random frequencies and the number of samples.

hidden_dim: The hidden dimension of each hidden layer.

num_layers: The number of hidden layers.

Examples

julia> FourierAttention(3, 1, sin; hidden_dims=10, num_layers=3,
                        modes=(1 => 10, 10 => 10, 50 => 10))
Chain(
    layer_1 = SkipConnection(
        FourierFeature(3 => 60),
        vcat
    ),
    layer_2 = PINNAttention(
        H_net = Dense(63 => 10, sin),   # 640 parameters
        U_net = Dense(63 => 10, sin),   # 640 parameters
        V_net = Dense(63 => 10, sin),   # 640 parameters
        fusion = TriplewiseFusion(
            layers = (layer_1 = Dense(10 => 10, sin), layer_2 = Dense(10 => 10, sin), layer_3 = Dense(10 => 10, sin), layer_4 = Dense(10 => 1)),  # 341 parameters
        ),
    ),
)         # Total: 2_261 parameters,
          #        plus 90 states, summarysize 176 bytes.

Sophon.FullyConnected — Method

FullyConnected(layer_dims::NTuple{N, Int}, activation; outermost = true, init_weight = kaiming_uniform)
FullyConnected(in_dims::Int, out_dims::Int, activation::Function;
               hidden_dims::Int, num_layers::Int, outermost=true, init_weight = kaiming_uniform)

Create fully connected layers.

Arguments

layer_dims: Number of dimensions of each layer.

hidden_dims: Number of hidden dimensions.

num_layers: Number of layers.

activation: Activation function.

Keyword Arguments

outermost: Whether to use activation function for the last layer. If false, the activation function is applied to the output of the last layer.

init_weight: Initialization method for the weights.

Example

julia> fc = FullyConnected((1, 12, 24, 32), relu)
Chain(
    layer_1 = Dense(1 => 12, relu),     # 24 parameters
    layer_2 = Dense(12 => 24, relu),    # 312 parameters
    layer_3 = Dense(24 => 32),          # 800 parameters
)         # Total: 1_136 parameters,
          #        plus 0 states, summarysize 48 bytes.

julia> fc = FullyConnected(1, 10, relu; hidden_dims=20, num_layers=3)
Chain(
    layer_1 = Dense(1 => 20, relu),     # 40 parameters
    layer_2 = Dense(20 => 20, relu),    # 420 parameters
    layer_3 = Dense(20 => 20, relu),    # 420 parameters
    layer_4 = Dense(20 => 10),          # 210 parameters
)         # Total: 1_090 parameters,
          #        plus 0 states, summarysize 64 bytes.

Sophon.MultiscaleFourier — Method

MultiscaleFourier(in_dims::Int, layer_dims::NTuple, activation=identity, modes::NTuple)

Multi-scale Fourier Feature Net.

x → FourierFeature → FullyConnected → y

Arguments

in_dims: The number of input dimensions.

layer_dims: A tuple of hidden dimensions used to construct FullyConnected.

activation: The activation function used to construct FullyConnected.

modes: A tuple of modes used to construct FourierFeature.

Keyword Arguments

modes: A tuple of modes used to construct FourierFeature.

Examples

julia> m = MultiscaleFourier(2, (30, 30, 1), sin; modes=(1 => 10, 10 => 10, 50 => 10))
Chain(
    layer_1 = FourierFeature(2 => 60),
    layer_2 = Dense(60 => 30, sin),     # 1_830 parameters
    layer_3 = Dense(30 => 30, sin),     # 930 parameters
    layer_4 = Dense(30 => 1),           # 31 parameters
)         # Total: 2_791 parameters,
          #        plus 60 states, summarysize 112 bytes.

References

[1] Wang, Sifan, Hanwen Wang, and Paris Perdikaris. “On the eigenvector bias of fourier feature networks: From regression to solving multi-scale pdes with physics-informed neural networks.” Computer Methods in Applied Mechanics and Engineering 384 (2021): 113938.

Sophon.Siren — Function

Siren(in_dims::Int, hidden_dim::Int, num_layers::Int; omega = 30f0)
Siren(layer_dims::NTuple{N, T}; omega = 30f0) where {N, T <: Int}

Sinusoidal Representation Network.

Keyword Arguments

omega: The ω₀ used for the first layer.

Examples

julia> Siren(3; hidden_dims=20, num_layers=3)
Chain(
    layer_1 = Sine(3 => 20),            # 80 parameters, plus 1
    layer_2 = Sine(20 => 20),           # 420 parameters
    layer_3 = Sine(20 => 20),           # 420 parameters
)         # Total: 920 parameters,
          #        plus 1 states, summarysize 292 bytes.

julia> Siren(3, 1; hidden_dims=20, num_layers=3)
Chain(
    layer_1 = Sine(3 => 20),            # 80 parameters, plus 1
    layer_2 = Sine(20 => 20),           # 420 parameters
    layer_3 = Sine(20 => 20),           # 420 parameters
    layer_4 = Sine(20 => 1),            # 21 parameters
)         # Total: 941 parameters,
          #        plus 1 states, summarysize 388 bytes.

References

[1] Sitzmann, Vincent, et al. "Implicit neural representations with periodic activation functions." Advances in Neural Information Processing Systems 33 (2020): 7462-7473.

Sophon.SirenAttention — Function

SirenAttention(in_dims::Int, out_dims::Int, activation::Function=sin;
hidden_dims::Int=512, num_layers::Int=6)

x -> Sine -> PINNAttention

Sophon.gaussian — Function

gaussian(x, a=0.2)

The Gaussian activation function.

\[e^{\frac{-0.5 x^{2}}{a^{2}}}\]

References

[1] Ramasinghe, Sameera, and Simon Lucey. "Beyond periodicity: Towards a unifying framework for activations in coordinate-mlps." arXiv preprint arXiv:2111.15135 (2021).

Sophon.kaiming_normal — Method

kaiming_normal(rng::AbstractRNG, size...; gain = √2f0)

Return an Array{Float32} of the given size containing random numbers taken from a normal distribution standard deviation gain / sqrt(fan_in)

References

[1] He, Kaiming, et al. "Delving deep into rectifiers: Surpassing human-level performance on imagenet classification." Proceedings of the IEEE international conference on computer vision. 2015.

Sophon.kaiming_uniform — Method

kaiming_uniform(rng::AbstractRNG, size...; gain = √2f0)

Return an Array{Float32} of the given size containing random numbers drawn from a uniform distribution on the interval [-x, x], where x = gain * sqrt(3/fan_in).

References

[1] He, Kaiming, et al. "Delving deep into rectifiers: Surpassing human-level performance on imagenet classification." Proceedings of the IEEE international conference on computer vision. 2015.

Sophon.wu — Function

wu(x,a=1)

An activation function I designed for use in coordinate-mlps.

\[\frac{x\left(5 x^{2}-1\right)}{\left(1+x^{2}\right)^{4}}\]