Sophon

Documentation for Sophon.

Sophon.FourierFeature

Sophon.PINNAttention

Sophon.Sine

Sophon.TriplewiseFusion

Sophon.FourierAttention

Sophon.FullyConnected

Sophon.MultiscaleFourier

Sophon.Siren

Sophon.FourierFeature — Type

FourierFeature(in_dims::Int, modes::NTuple{N,Pair{S,T}}) where {N,S,T<:Int}

Fourier Feature Network.

Arguments

in_dims: Input dimension.

modes: A tuple of pairs of std => out_dims, where std is the standard deviation of the Gaussian distribution, and out_dims is the output dimension.

Inputs

x: An an AbstractArray with size(x, 1) == in_dims.

Returns

An AbstractArray with size(y, 1) == sum(last(modes) * 2).

The states of the layers.

States

modes: Random frequencies.

Examples

FourierFeature(2, (1 => 3, 50 => 4))

References

[1] Tancik, Matthew, et al. “Fourier features let networks learn high frequency functions in low dimensional domains.” Advances in Neural Information Processing Systems 33 (2020): 7537-7547.

Sophon.PINNAttention — Type

PINNAttention(H_net, U_net, V_net, fusion_layers)
PINNAttention(in_dims::Int, hidden_dim::Int, num_layers::Int, activation)

The output dimesion of H_net and the input dimension of fusion_layers must be the same. For the second and the third constructor, Dense layers is used for H_net, U_net, and V_net.

                 x → U_net → u                           u
                               ↘                           ↘
x → H_net →  h1 → fusionlayer1 → connection → fusionlayer2 → connection
                               ↗                           ↗
                 x → V_net → v                           v

Arguments

- `H_net`: `AbstractExplicitLayer`
- `U_net`: `AbstractExplicitLayer`
- `V_net`: `AbstractExplicitLayer`
- `in_dims`: The input dimension.
- `hidden_dims`: The output dimension of `H_net`.
- `fusion_layers`: `AbstractExplicitLayer` or a tuple of integeters. In the latter case,
    fully connected layers are used.

References

[1] Wang, Sifan, Yujun Teng, and Paris Perdikaris. "Understanding and mitigating gradient flow pathologies in physics-informed neural networks." SIAM Journal on Scientific Computing 43.5 (2021): A3055-A3081

Sophon.Sine — Type

Sine(in_dims::Int, out_dims::Int, activation=sin; is_first::Bool = false, omega::AbstractFloat = 30f0)

Sinusoidal layer.

Example

s = Sine(2, 2; is_first=true) # first layer
s = Sine(2, 2) # hidden layer
s = Sine(2, 2, identity) # last layer

Sophon.TriplewiseFusion — Type

TriplewiseFusion(connection, layers...)

         u1                    u2
            ↘                     ↘
h1 → layer1 → connection → layer2 → connection
            ↗                     ↗
         v1                    v2

Arguments

connection: Takes 3 inputs and combines them

layers: AbstractExplicitLayers or a Chain.

Inputs

Layer behaves differently based on input type:

A tripe of (h, u, v), where u and v itself are tuples of length N, the layers is also a tuple of length N. The computation is as follows

for i in 1:N
    h = connection(layers[i](h), u[i], v[i])
end

A triple of (h, u, v), where u and v are AbstractArrays.

for i in 1:N
    h = connection(layers[i](h), u, v)
end

Returns

See Inputs section for how the return value is computed

Updated model state for all the contained layers

Parameters

Parameters of each layer wrapped in a NamedTuple with fields = layer_1, layer_2, ..., layer_N

States

States of each layer wrapped in a NamedTuple with fields = layer_1, layer_2, ..., layer_N

Sophon.FourierAttention — Method

FourierAttention(in_dims::Int, hidden_dim::Int, num_layers::Int, activation; modes)

x → [FourierFeature(x); x] → PINNAttention

Arguments

in_dims: The input dimension.

hidden_dim: The hidden dimension of each hidden layer.

num_layers: The number of hidden layers.

Keyword Arguments

modes: A tuple of pairs of random frequencies and the number of samples.

Sophon.FullyConnected — Method

FullyConnected(in_dims, hidden_dims::NTuple{N, Int}, activation; outermost = false)
FullyConnected(in_dims, hidden_dims, num_layers, activation; outermost = false)

Create fully connected layers.

Arguments

in_dims: Input dimension.

hidden_dims: Hidden dimensions.

num_layers: Number of layers.

activation: Activation function.

Keyword Arguments

outermost: Whether to use activation function for the last layer.

Sophon.MultiscaleFourier — Method

MultiscaleFourier(in_dims::Int, out_dims::NTuple, activation=identity, modes::NTuple)

Multi-scale Fourier Feature Net.

x → FourierFeature → FullyConnected → y

Arguments

in_dims: The number of input dimensions.

out_dims: A tuple of output dimensions used to construct FullyConnected.

activation: The activation function used to construct FullyConnected.

modes: A tuple of modes used to construct FourierFeature.

References

[1] Wang, Sifan, Hanwen Wang, and Paris Perdikaris. “On the eigenvector bias of fourier feature networks: From regression to solving multi-scale pdes with physics-informed neural networks.” Computer Methods in Applied Mechanics and Engineering 384 (2021): 113938.

Sophon.Siren — Method

Siren(in_dims::Int, hidden_dim::Int, num_layers::Int; omega = 30f0)
Siren(in_dims::Int, hidden_dims::NTuple{N, T}; omega = 30f0) where {N, T <: Int}

Sinusoidal Representation Network.

Keyword Arguments

omega: The ω₀ used for the first layer.

References

[1] Sitzmann, Vincent, et al. "Implicit neural representations with periodic activation functions." Advances in Neural Information Processing Systems 33 (2020): 7462-7473.