Sophon

Documentation for Sophon.

Sophon.DeepONet

Sophon.DiscreteFourierFeature

Sophon.FourierFeature

Sophon.NonAdaptiveTraining

Sophon.PINN

Sophon.PINNAttention

Sophon.QuasiRandomSampler

Sophon.RBF

Sophon.TriplewiseFusion

Sophon.BACON

Sophon.FourierAttention

Sophon.FourierFilterNet

Sophon.FourierNet

Sophon.FullyConnected

Sophon.Sine

Sophon.Siren

Sophon.gaussian

Sophon.DeepONet — Type

DeepONet(branch_net, trunk_net;
         flatten_layer=FlattenLayer(),
         linear_layer=NoOpLayer(),
         bias=Scalar())
DeepONet(layer_sizes_branch, activation_branch,
         layer_sizes_trunk,
         activation_trunk,
         layer_sizes_linear=nothing)

Deep operator network. Note that the branch net supports multi-dimensional inputs. The flatten_layer flatten the output of the branch net to a matrix, and the linear_layer is applied to the flattened. In this case, linear_layer must be given to transform the flattened matrix to the correct shape.

v → branch_net → flatten_layer → linear_layer → b
                                                  ↘
                                                    b' * t + bias → u
                                                  ↗
                                ξ → trunk_net → t

Arguments

branch_net: The branch net.

trunk_net: The trunk net.

Keyword Arguments

flatten_layer: The layer to flatten a multi-dimensional array to a matrix.

linear_layer: The layer to apply a linear transformation to the output of the flatten_layer.

Inputs

(v, ξ): v is an array of shape $(b_1,b_2,...b_d, m)$, where d is the dimension of the input function, and m is the number of input functions. ξ is a matrix of shape $(d', n)$, where $d'$ is the dimension of the output function, and m is the number of "sensors".

Returns

A matrix of shape $(m, n)$.

Examples

julia> deeponet = DeepONet((3, 5, 4), relu, (2, 6, 4, 4), tanh)
DeepONet(
    branch_net = Chain(
        layer_1 = Dense(3 => 5, relu),  # 20 parameters
        layer_2 = Dense(5 => 4),        # 24 parameters
    ),
    trunk_net = Chain(
        layer_1 = Dense(2 => 6, tanh_fast),  # 18 parameters
        layer_2 = Dense(6 => 4, tanh_fast),  # 28 parameters
        layer_3 = Dense(4 => 4, tanh_fast),  # 20 parameters
    ),
    flatten_layer = FlattenLayer(),
    linear_layer = NoOpLayer(),
    bias = Scalar(),                    # 1 parameters
)         # Total: 111 parameters,
          #        plus 0 states, summarysize 80 bytes.

Reference

Lu Lu, Pengzhan Jin, Guofei Pang, Zhongqiang Zhang, George Em Karniadakis (2021)

Sophon.DiscreteFourierFeature — Type

DiscreteFourierFeature(in_dims::Int, out_dims::Int, N::Int, period::Real)

The discrete Fourier filter proposed in David B. Lindell, Dave Van Veen, Jeong Joon Park, Gordon Wetzstein (2021). For a periodic function with period $P$, the Fourier series in amplitude-phase form is

\[s_N(x)=\frac{a_0}{2}+\sum_{n=1}^N{a_n}\cdot \sin \left( \frac{2\pi}{P}nx+\varphi _n \right)\]

The output is guaranteed to be periodic.

Arguments

in_dims: Number of the input dimensions.

out_dims: Number of the output dimensions.

N: $N$ in the formula.

period: $P$ in the formula.

Sophon.FourierFeature — Type

FourierFeature(in_dims::Int, std::NTuple{N,Pair{S,T}}) where {N,S,T<:Int}
FourierFeature(in_dims::Int, frequencies::NTuple{N, T}) where {N, T <: Real}
FourierFeature(in_dims::Int, out_dims::Int, std::Real)

Fourier Feature Network.

Arguments

in_dims: Number of the input dimensions.

std: A tuple of pairs of sigma => out_dims, where sigma is the standard deviation of the Gaussian distribution.

\[\phi^{(i)}(x)=\left[\sin \left(2 \pi W^{(i)} x\right) ; \cos 2 \pi W^{(i)} x\right],\ W^{(i)} \sim \mathcal{N}\left(0, \sigma^{(i)}\right),\ i\in 1, \dots, D\]

frequencies: A tuple of frequencies (f1,f2,...,fn).

\[\phi^{(i)}(x)=\left[\sin \left(2 \pi f_i x\right) ; \cos 2 \pi f_i x\right]\]

Inputs

x: AbstractArraywithsize(x, 1) == in_dims`.

Returns

AbstractArray with size(y, 1) == sum(last(modes) * 2).

States

The weight W in case 1, otherwise NamedTuple().

Examples

julia> f = FourierFeature(2,10,1) # Random Fourier Feature
FourierFeature(2 => 10)

julia> f = FourierFeature(2, (1 => 3, 50 => 4)) # Multi-scale Random Fourier Features
FourierFeature(2 => 14)

julia>  f = FourierFeature(2, (1,2,3,4)) # Predefined frequencies
FourierFeature(2 => 16)

References

Ali Rahimi, Benjamin Recht (2007)

Matthew Tancik, Pratul Srinivasan, Ben Mildenhall, Sara Fridovich-Keil, Nithin Raghavan, Utkarsh Singhal, Ravi Ramamoorthi, Jonathan Barron, Ren Ng (2020)

Sifan Wang, Hanwen Wang, Paris Perdikaris (2021)

Sophon.NonAdaptiveTraining — Type

NonAdaptiveTraining(pde_weights=1, bcs_weights=pde_weights)

Fixed weights for the loss functions.

Sophon.PINN — Type

PINN(chain; device_type::Type=Array{Float64})

Arguments

chain: AbstractExplicitLayer or a named tuple of AbstractExplicitLayers.

device_type: Array{T} or CuArray{T}, or any other array types.

Sophon.PINNAttention — Type

PINNAttention(H_net, U_net, V_net, fusion_layers)
PINNAttention(in_dims::Int, out_dims::Int, activation::Function=sin;
              hidden_dims::Int, num_layers::Int)

The output dimesion of H_net and the input dimension of fusion_layers must be the same. For the second and the third constructor, Dense layers is used for H_net, U_net, and V_net. Note that the first constructer does not contain the output layer.

                 x → U_net → u                           u
                               ↘                           ↘
x → H_net →  h1 → fusionlayer1 → connection → fusionlayer2 → connection
                               ↗                           ↗
                 x → V_net → v                           v

Arguments

H_net: AbstractExplicitLayer.

U_net: AbstractExplicitLayer.

V_net: AbstractExplicitLayer.

fusion_layers: Chain.

Keyword Arguments

num_layers: The number of hidden layers.

hidden_dims: The number of hidden dimensions of each hidden layer.

Reference

Sifan Wang, Yujun Teng, Paris Perdikaris (2021)

Sophon.QuasiRandomSampler — Type

QuasiRandomSampler(pde_points, bcs_points=pde_points;
                   device_type::Type=Array{Float64}
                   sampling_alg=LatinHypercubeSample())

Sampler to generate the datasets for PDE and boundary conditions using a quisa-random sampling algorithm. It momerizes the domain of the PDE and the boundary conditions, and you can call sample on it to generate the datasets.

Sophon.RBF — Type

RBF(in_dims::Int, out_dims::Int, num_centers::Int=out_dims; sigma::AbstractFloat=0.2f0)

Normalized Radial Basis Fuction Network.

Sophon.TriplewiseFusion — Type

TriplewiseFusion(connection, layers...)

     u1                    u2
        ↘                     ↘
h1 → layer1 → connection → layer2 → connection
        ↗                     ↗
     v1                    v2

Arguments

connection: A functio takes 3 inputs and combines them.

layers: AbstractExplicitLayers or a Chain.

Inputs

Layer behaves differently based on input type:

A tripe of (h, u, v), where u and v itself are tuples of length N, the layers is also a tuple of length N. The computation is as follows

for i in 1:N
    h = connection(layers[i](h), u[i], v[i])
end

A triple of (h, u, v), where u and v are AbstractArrays.

for i in 1:N
    h = connection(layers[i](h), u, v)
end

Parameters

Parameters of each layer wrapped in a NamedTuple with fields = layer_1, layer_2, ..., layer_N

States

States of each layer wrapped in a NamedTuple with fields = layer_1, layer_2, ..., layer_N

Sophon.BACON — Method

BACON(in_dims::Int, out_dims::Int, N::Int, period::Real; hidden_dims::Int, num_layers::Int)

Band-limited Coordinate Networks (BACON) from David B. Lindell, Dave Van Veen, Jeong Joon Park, Gordon Wetzstein (2021). Similar to FourierFilterNet but the frequcies are dicrete and nontrainable.

Tips: It is recommended to set period to be 1,2,π or 2π for better performance.

Sophon.FourierAttention — Function

FourierAttention(in_dims::Int, out_dims::Int, activation::Function=sin;
                 hidden_dims::Int=512, num_layers::Int=6, modes::NTuple)

x → [FourierFeature(x); x] → PINNAttention

Arguments

in_dims: The input dimension.

Keyword Arguments

modes: A tuple of pairs of random frequencies and the number of samples.

hidden_dim: The hidden dimension of each hidden layer.

num_layers: The number of hidden layers.

Examples

julia> FourierAttention(3, 1, sin; hidden_dims=10, num_layers=3,
                        modes=(1 => 10, 10 => 10, 50 => 10))
Chain(
    layer_1 = SkipConnection(
        FourierFeature(3 => 60),
        vcat
    ),
    layer_2 = PINNAttention(
        H_net = Dense(63 => 10, sin),   # 640 parameters
        U_net = Dense(63 => 10, sin),   # 640 parameters
        V_net = Dense(63 => 10, sin),   # 640 parameters
        fusion = TriplewiseFusion(
            layers = (layer_1 = Dense(10 => 10, sin), layer_2 = Dense(10 => 10, sin), layer_3 = Dense(10 => 10, sin), layer_4 = Dense(10 => 1)),  # 341 parameters
        ),
    ),
)         # Total: 2_261 parameters,
          #        plus 90 states, summarysize 176 bytes.

Sophon.FourierFilterNet — Method

FourierFilterNet(in_dims::Int, out_dims::Int; hidden_dims::Int, num_layers::Int,
                 bandwidth::Real)

Keyword Arguments

bandwidth: The maximum bandwidth of the network. The bandwidth is the sum of each filter's bandwidth.

Parameters

Parameters of the filters:

\[ W\sim \mathcal{U}(-\frac{ω}{n}, \frac{ω}{n}), \quad b\sim \mathcal{U}(-\pi, \pi),\]

where n is the number of filters.

For a periodic function with period $P$, the Fourier series in amplitude-phase form is

\[s_N(x)=\frac{a_0}{2}+\sum_{n=1}^N{a_n}\cdot \sin \left( \frac{2\pi}{P}nx+\varphi _n \right)\]

We have the following relation between the banthwidth and the parameters of the model:

\[ω = 2πB=\frac{2πN}{P}.\]

where $B$ is the bandwidth of the network.

References

Rizal Fathony, Anit Kumar Sahu, Devin Willmott, J Zico Kolter (2021)

David B. Lindell, Dave Van Veen, Jeong Joon Park, Gordon Wetzstein (2021)

Sophon.FourierNet — Method

FourierNet(ayer_sizes::NTuple, activation, modes::NTuple)

A model that combines FourierFeature and FullyConnected.

x → FourierFeature → FullyConnected → y

Arguments

in_dims: The number of input dimensions.

layer_sizes: A tuple of hidden dimensions used to construct FullyConnected.

activation: The activation function used to construct FullyConnected.

modes: A tuple of modes used to construct FourierFeature.

Examples

julia> FourierNet((2, 30, 30, 1), sin, (1 => 10, 10 => 10, 50 => 10))
Chain(
    layer_1 = FourierFeature(2 => 60),
    layer_2 = Dense(60 => 30, sin),     # 1_830 parameters
    layer_3 = Dense(30 => 30, sin),     # 930 parameters
    layer_4 = Dense(30 => 1),           # 31 parameters
)         # Total: 2_791 parameters,
          #        plus 60 states, summarysize 112 bytes.

julia> FourierNet((2, 30, 30, 1), sin, (1, 2, 3, 4))
Chain(
    layer_1 = FourierFeature(2 => 16),
    layer_2 = Dense(16 => 30, sin),     # 510 parameters
    layer_3 = Dense(30 => 30, sin),     # 930 parameters
    layer_4 = Dense(30 => 1),           # 31 parameters
)         # Total: 1_471 parameters,
          #        plus 4 states, summarysize 96 bytes.

Sophon.FullyConnected — Method

FullyConnected(layer_sizes::NTuple{N, Int}, activation; outermost = true,
               init_weight = kaiming_uniform(activation),
               init_bias = zeros32)
FullyConnected(in_dims::Int, out_dims::Int, activation::Function;
               hidden_dims::Int, num_layers::Int, outermost=true,
               init_weight = kaiming_uniform(activation),
               init_bias = zeros32)

Create fully connected layers.

Arguments

layer_sizes: Number of dimensions of each layer.

hidden_dims: Number of hidden dimensions.

num_layers: Number of layers.

activation: Activation function.

Keyword Arguments

outermost: Whether to use activation function for the last layer. If false, the activation function is applied to the output of the last layer.

init_weight: Initialization method for the weights.

Example

julia> fc = FullyConnected((1, 12, 24, 32), relu)
Chain(
    layer_1 = Dense(1 => 12, relu),     # 24 parameters
    layer_2 = Dense(12 => 24, relu),    # 312 parameters
    layer_3 = Dense(24 => 32),          # 800 parameters
)         # Total: 1_136 parameters,
          #        plus 0 states, summarysize 48 bytes.

julia> fc = FullyConnected(1, 10, relu; hidden_dims=20, num_layers=3)
Chain(
    layer_1 = Dense(1 => 20, relu),     # 40 parameters
    layer_2 = Dense(20 => 20, relu),    # 420 parameters
    layer_3 = Dense(20 => 20, relu),    # 420 parameters
    layer_4 = Dense(20 => 10),          # 210 parameters
)         # Total: 1_090 parameters,
          #        plus 0 states, summarysize 64 bytes.

Sophon.Sine — Method

Sine(in_dims::Int, out_dims::Int; omega::Real)

Sinusoidal layer.

Example

s = Sine(2, 2; omega=30.0f0) # first layer
s = Sine(2, 2) # hidden layer

Sophon.Siren — Method

Siren(in_dims::Int, out_dims::Int; hidden_dims::Int, num_layers::Int, omega=30.0f0,
      init_weight=nothing))
Siren(layer_sizes::Int...; omega=30.0f0, init_weight=nothing)

Sinusoidal Representation Network.

Keyword Arguments

omega: The ω₀ used for the first layer.

init_weight: The initialization algorithm for the weights of the input layer. Note that all hidden layers use kaiming_uniform as the initialization algorithm. The default is
\[ W\sim \mathcal{U}\left(-\frac{\omega}{fan_{in}}, \frac{\omega}{fan_{in}}\right)\]

Examples

julia> Siren(2, 32, 32, 1; omega=5.0f0)
Chain(
    layer_1 = Dense(2 => 32, sin),      # 96 parameters
    layer_2 = Dense(32 => 32, sin),     # 1_056 parameters
    layer_3 = Dense(32 => 1),           # 33 parameters
)         # Total: 1_185 parameters,
          #        plus 0 states, summarysize 48 bytes.

julia> Siren(3, 1; hidden_dims=20, num_layers=3)
Chain(
    layer_1 = Dense(3 => 20, sin),      # 80 parameters
    layer_2 = Dense(20 => 20, sin),     # 420 parameters
    layer_3 = Dense(20 => 20, sin),     # 420 parameters
    layer_4 = Dense(20 => 1),           # 21 parameters
)         # Total: 941 parameters,
          #        plus 0 states, summarysize 64 bytes.

# Use your own initialization algorithm for the input layer.
julia> init_weight(rng::AbstractRNG, out_dims::Int, in_dims::Int) = randn(rng, Float32, out_dims, in_dims) .* 2.5f0
julia> chain = Siren(2, 1; num_layers = 4, hidden_dims = 50, init_weight = init_weight)

Reference

Vincent Sitzmann, Julien Martel, Alexander Bergman, David Lindell, Gordon Wetzstein (2020)

Sophon.gaussian — Function

gaussian(x, a=0.2)

The Gaussian activation function.

\[e^{\frac{- x^{2}}{2a^{2}}}\]

Reference

Sameera Ramasinghe, Simon Lucey (2021)